# chaoscenter/authentication/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Chaoscenter API documentation",
        "contact": {}
    },
    "paths": {
...
        "/readiness": {
            "get": {
                "description": "Return list of tags.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MiscRouter"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrServerError"
                        }
                    }
                }
            }
        },
...
        "/status": {
            "get": {
                "description": "Status will request users list and return, if successful, a http code 200.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MiscRouter"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "The authorization server encountered an unexpected condition that prevented it from fulfilling the request"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string"
                }
            }
        },
...
    }
}